Fix broken downvote menu on bad comments (#432)
* Fix broken downvote menu on bad comments

The reduced opacity on a comment with the 'bad'-class affects the
stacking order of its child elements and breaks the downvote menu. To
remedy this, the downvote menu is placed outside of the non-opaque
comment element.

* Increase z-index of #downvote_why

The lower end of the comment icon in the mobile style sheet had a higher
z-index and ended up in front of the downvote menu.